Enhancing troops movements
Added the option to move ground and naval troops through cities.
I'm not a programmer, but I really liked the project, and I'm trying to contribute in the best way possible.
In this contribution, I added a file called sendTroopsShips.py divided into functions.
First, it retrieves the information from the HTML file under "Troops in the town" for units/ships, summing up all the information received from the cities and dividing them into Units and Ships. The next step is to select the source and destination information for troop movement.
I know there are some improvements that can be made, such as travel time, sending to foreign cities, the ability to choose specific units for movement, etc. However, these improvements will be implemented in other commits.