Add loop option, bugfix, add status updates
Added 3rd option for investigation, automatic experiment conduction. Instead of setting a fixed amount of experiments to conduct based on the available resources, this adds an option to check the selected town every 4 hours and conduct an experiment if there are enough resources.

Added a cooldown check before trying to conduct an experiment.

Switched most Telegram notifications to status updates.