Remove gettext from project
Gettext is unnecessary as ikabot is a relatively small project. We do not have enough resources to add different locales and even if we did, it would require a lot of resources to maintain those localizations when new modules are written in english.

To add to this, gettext wasn't even used across the entire project, some modules used it for all strings, some modules used it only for some strings, and not other and some modules didn't use it at all.

All in all, I think it's safe to assume ikabot is better off without gettext