Split module execution from module configuration
Now every module in the function directory has a function with the same name as the module and a do_it function. The function with the same name is the entry function and it's supposed to be used as a configurator. It returns a dictionary containing the parameters for the do_it function. The do_it function is supposed to run in the background as a task.