Multi-city option for consolidateResources (#335)
* New helper function

Add a new helper function to prompt the user to select cities to ignore, returning a list of user's cities.

* Multi-city option for consolidateResources

Ability to use multiple cities and general optimization for the status messaging. Also added an optional message for the helper

* Remove bot message

Removed the bot message for no excess resources to send information, after waking up to 56 Telegram messages lol.